Fixed ID data type mismatch in LogincheckRepository from Long to Integer

+++ DevFolio/ResumePortal/ResumePortal/src/main/java/com/Prathamesh/ResumePortal/repository/LogincheckRepository.java
@@ -4,6 +4,6 @@ import com.Prathamesh.ResumePortal.model.LogincheckModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface LogincheckRepository extends JpaRepository<LogincheckModel, Long> {
+public interface LogincheckRepository extends JpaRepository<LogincheckModel, Integer> {
     LogincheckModel findByUsernameAndPassword(String username, String password);
 }
